+++ NEWS.md
@@ -1,10 +1,14 @@
+# plotBart 0.1.23
+- added `plot_trank()` function
+- added adition paramaters to `plot_trace()`
+
 # plotBart 0.1.22
 - minor bug fix
 
 # plotBart 0.1.21
 - Added `view_overlap` argument to `plot_SATE()`function. 
 - Allows user to view what results would have been under alternative common support rules. 
-- Improvments to `plot_common_support()`
+- Improvements to `plot_common_support()`
 - Added `trim` feature to `plot_overlap_pScores()` this allows users to zoom in on relevant area of the plot for overlap checks
 
 # plotBart 0.1.20